# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  "/lang/{lang}":
    get:
      tags:
        - Miscellaneous
      description: returns the JSON language pack given the language code
      operationId: getI18nLang
      parameters:
        - in: path
          name: lang
          required: true
          description: JSON language pack required
          schema:
            type: string
      responses:
        "200":
          description: requested language pack
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/LanguagePack"
...
  /settings:
    get:
      tags:
        - Settings
      description: returns settings from DB
      operationId: getSettings
      responses:
        "200":
          description: settings object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Settings"

    put:
      tags:
        - Settings
      description: returns updated settings from the DB.
      operationId: updateSettings
      requestBody:
        description: updated settings field values
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: updated settings object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  /settings/smtp/test:
    post:
      tags:
        - Settings
      description: test smtp settings
      operationId: testSMTPSettings
      requestBody:
        description: updated SMTP settings field values
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SMTPTest"
      responses:
        "200":
          description: updated SMTP test settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
...
  /subscribers:
    get:
      tags:
        - Subscribers
      description: returns all subscribers.
      operationId: getSubscribers
      parameters:
        - in: query
          name: page
          description: Page number for paginated results.
          required: false
          schema:
            type: integer
            format: int32
        - in: query
          name: per_page
          description: Number of items per page. Use an integer for specific page size or 'all' to retrieve all results
          required: false
          schema:
            oneOf:
              - type: integer
                description: Number of items to return per page
              - type: string
                enum: ["all"]
                description: Return all results without pagination
        - in: query
          name: query
          description: query subscribers with an SQL expression.
          required: false
          schema:
            type: string
        - in: query
          name: order_by
          description: Result sorting field. Options are name, status, created_at, updated_at
          required: false
          schema:
            type: string
            enum: ["name", "status", "created_at", "updated_at"]
        - in: query
          name: order
          description: ASC|DESC Sort by ascending or descending order.
          required: false
          schema:
            type: string
            enum: ["ASC", "DESC"]
        - in: query
          name: subscription_status
          description: Subscription status to filter by if there are one or more list_ids.
          required: false
          schema:
            type: string
        - in: query
          name: list_id
          description: ID of lists to filter by. Repeat in the query for multiple values.
          required: false
          schema:
            type: array
            items:
              type: integer

      responses:
        "200":
          description: subscribers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/Subscriber"
                      query:
                        type: string
                      total:
                        type: integer
                      per_page:
                        type: integer
                      page:
                        type: integer

    post:
      description: handles creation of new subscriber
      operationId: createSubscriber
      tags:
        - Subscribers
      requestBody:
        description: new subscriber info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSubscriber"

      responses:
        "200":
          description: subscriber object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Subscriber"

    delete:
      description: handles subscribers deletion
      operationId: deleteSubscriberByList
      tags:
        - Subscribers
      parameters:
        - in: query
          name: id
          required: true
          description: subscriber id/s to be deleted
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/{id}":
    get:
      description: handles the retrieval of a single subscriber by ID.
      operationId: getSubscriberById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the subscriber you want to get.
          schema:
            type: integer

      responses:
        "200":
          description: gets a single subscriber.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Subscriber"

    put:
      description: modify subscriber data
      operationId: updateSubscriberById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: The id of subscriber to update
          schema:
            type: integer
      requestBody:
        description: new subscriber info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriber"
      responses:
        "200":
          description: returns updated subscriber.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Subscriber"

    delete:
      description: handles subscriber deletion based on id
      operationId: deleteSubscriberById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the subscriber you want to get.
          schema:
            type: integer

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  /subscribers/lists:
    put:
      description: handles bulk addition or removal of subscribers
      operationId: manageSubscriberLists
      tags:
        - Subscribers
      requestBody:
        description: The list of subscribers details to add or remove
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/lists/{id}":
    put:
      description: handles bulk addition or removal of subscribers for a specified list id
      operationId: manageSubscriberListById
      tags:
        - Subscribers
      requestBody:
        description: The list of subscribers to add or remove
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"
      parameters:
        - in: path
          name: id
          required: true
          description: The id of list you want to update
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/blocklist":
    put:
      description: handles blocklisting of subscriber list
      operationId: manageBlocklistBySubscriberList
      tags:
        - Subscribers
      requestBody:
        description: The list of subscribers to blocklist
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/{id}/blocklist":
    put:
      description: handles the blocklisting of one or more subscribers.
      operationId: manageBlocklistSubscribersById
      tags:
        - Subscribers
      requestBody:
        description: The id of subscriber to add or remove
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of the subscriber you want to blocklist.
          schema:
            type: integer
            format: int32

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/{id}/export":
    get:
      description: retrieves a subscriber's profile
      operationId: exportSubscriberDataByID
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: The id value of subscriber profile you want to export
          schema:
            type: integer
      responses:
        "200":
          headers:
            Cache-Control:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="data.json"

          description: subscriber data object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberData"
  "/subscribers/{id}/bounces":
    get:
      description: retrieves a subscriber's bounce records
      operationId: getSubscriberBouncesById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: subscriber id
          schema:
            type: integer
      responses:
        "200":
          description: list of bounce records of a subscriber
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bounce"

    delete:
      description: deletes a subscriber's bounce records
      operationId: deleteSubscriberBouncesById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: subscriber id
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/{id}/optin":
    post:
      description: sends an optin confirmation e-mail to a subscriber.
      operationId: subscriberSendOptinById
      tags:
        - Subscribers
      parameters:
        - in: path
          name: id
          required: true
          description: sends an optin confirmation e-mail to a subscriber
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/query/delete":
    post:
      description: bulk deletes based on an arbitrary SQL expression.
      operationId: deleteSubscriberByQuery
      tags:
        - Subscribers
      requestBody:
        description: Arbitrary SQL expression.
        content:
          text/plain:
            schema:
              type: string
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/query/blocklist":
    put:
      description: bulk blocklists subscribers based on an arbitrary SQL expression.
      operationId: blocklistSubscribersQuery
      tags:
        - Subscribers
      requestBody:
        description: Arbitrary SQL expression.
        content:
          text/plain:
            schema:
              type: string
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  "/subscribers/query/lists":
    put:
      description: bulk adds/removes/unsubscribes subscribers from one or more lists based on an arbitrary SQL expression.
      operationId: manageSubscriberListsByQuery
      tags:
        - Subscribers
      requestBody:
        description: Arbitrary SQL expression.
        content:
          text/plain:
            schema:
              type: string
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriberQueryRequest"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
...
components:
  schemas:
...
    SMTPTest:
      type: object
      properties:
        uuid:
          type: string
        enabled:
          type: boolean
        host:
          type: string
        hello_hostname:
          type: string
        port:
          type: integer
        auth_protocol:
          type: string
        username:
          type: string
        email_headers:
          type: array
          items:
            type: object
        max_conns:
          type: integer
        max_msg_retries:
          type: integer
        idle_timeout:
          type: string
        wait_timeout:
          type: string
        tls_type:
          type: string
        tls_skip_verify:
          type: boolean
        strEmailHeaders:
          type: string
        password:
          type: string
        email:
          type: string
...
    Subscriber:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        uuid:
          type: string
        email:
          type: string
        name:
          type: string
        attribs:
          type: object
          additionalProperties: true
        status:
          type: string
        lists:
          type: array
          items:
            type: object
            properties:
              subscription_status:
                type: string
              id:
                type: integer
              uuid:
                type: string
              name:
                type: string
              type:
                type: string
              tags:
                type: array
                items:
                  type: string
              created_at:
                type: string
              updated_at:
                type: string
...